requirements.txt
fastapi
uvicorn
python-multipart
pillow
ultralytics
torch==2.1.2+cpu
torchvision==0.16.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
